Support exporting data as .csv via chrome.fileSystem API
Provides a button on the results page that will be active after App’s
this.state.questionId > -1 and voting is not currently in progress,
i.e. at least one vote has taken place this session.

The user will be prompted to save file to a location and notified via
chrome.notifications API notification whether or not the export
succeeded.

This includes the addition of utils.js and a chrome.notifcations API
wrapper function for simplicity/brevity of calls.